refactor(senpi-task): extract the task record writer into record-write
record-store.ts sits at the 250 pure-LOC ceiling, so the atomic record write moves into its own module before it grows a Windows rename retry (#8050). Pure move: writeRecord now takes the resolved record path and TaskRecordCollisionError is re-exported from record-store so every import path stays valid.
